requirements.txt
python-pptx==0.6.23
openpyxl==3.1.2
openai>=1.99.5
click==8.1.7
python-dotenv>=1.1.0
Pillow>=10.0.0
svglib>=1.6.0
reportlab>=4.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0

